Improved security of backup script
Backup script tgz's folders, than encrypts them with GPG, uploads it via
ftps and then deletes all the files.

The problem is, the plaintext .tgz (not yet encrypted) file is
accessible in /tmp until ftps upload finished. This is severe security
risk, as there is pretty large attack window.

Now, the plaintext .tgz is deleted as soon as it's encrypted with GPG.

For even more privacy and security (which I as sinlge user of my
computer don't need), change the /tmp to your home directory so no
non-root user can access it anyway.