Fixed root taking over permissions over pazi DB file
I once used root with zsh and pazi (with the exactly same
shell setup), and did cd. This made a write as root to pazi
db file, which turned it's permission to root.

Because when pazi updates db file, it does so atomically.
This means that pazi creates new db file (like `.new`), then
renames old one (like `.bak`) and then renames `.new` to the
normal DB name, and deletes `.bak`. Note that I made up those
extensions to explain what it does.

This caused pazi to break with permission denied whenever it
tried to do anything with it's DB.

This fixes it, as it sets $HOME to /root/ (in case of root user),
even when we have the -E flag. So pazi should save what root
does to /root/.config/pazi directory, instead of home directory
of user who invoked su(do).